Fixed FIDE and FFE downloads

+++ src/database/sqlite/ffe_database.py
@@ -181,7 +181,8 @@ class FfeDatabase(SQLiteDatabase):
             if save:
                 save.rename(self.file)
             return False
-        save.unlink(missing_ok=True)
+        if save:
+            save.unlink(missing_ok=True)
         print_interactive_success(
             _('{number} players written.').format(number=players_number)
         )
